Move FFNN and LSTM definitions into their own file
