+++ cipherTypeDetection/models/ffnn.py
@@ -1,0 +1,24 @@
+import torch
+from torch import nn
+import torch.nn.functional as F
+
+
+class FFNN(nn.Module):
+    def __init__(self, input_size, hidden_size, output_size, num_hidden_layers):
+        super().__init__()
+
+        # saves parameters so that they can be saved and loaded later
+        self.input_size = input_size
+        self.hidden_size = hidden_size
+        self.output_size = output_size
+        self.num_hidden_layers = num_hidden_layers
+
+        layers = [nn.Linear(input_size, hidden_size), nn.ReLU()]
+        for _ in range(num_hidden_layers - 1):
+            layers += [nn.Linear(hidden_size, hidden_size), nn.ReLU()]
+        layers.append(nn.Linear(hidden_size, output_size))
+        self.net = nn.Sequential(*layers)
+
+    def forward(self, x):
+        return self.net(x)
+
